Initial Implementation of LineKickPlanner
