+++ soccer/gameplay/robocup-py.cpp
@@ -622,6 +622,7 @@ BOOST_PYTHON_MODULE(robocup) {
         .def("set_world_vel", &OurRobot::worldVelocity)
         .def("face", &OurRobot::face)
         .def("pivot", &OurRobot::pivot)
+        .def("line_kick", &OurRobot::lineKick)
         .def("set_planning_priority", &OurRobot::setPlanningPriority)
         .def("set_max_angle_speed", OurRobot_set_max_angle_speed)
         .def("set_max_speed", OurRobot_set_max_speed)
