+++ soccer/planning/LineKickPlanner.hpp
@@ -2,7 +2,7 @@
 
 #include "SingleRobotPathPlanner.hpp"
 #include <Geometry2d/Point.hpp>
-
+#include "RRTPlanner.hpp"
 class Configuration;
 class ConfigDouble;
 
@@ -10,7 +10,7 @@ namespace Planning {
 
 class LineKickPlanner : public SingleRobotPathPlanner {
 public:
-    LineKickPlanner() : SingleRobotPathPlanner(false){};
+    LineKickPlanner() : SingleRobotPathPlanner(false), rrtPlanner(250){};
     virtual std::unique_ptr<Path> run(SinglePlanRequest& planRequest) override;
 
     virtual MotionCommand::CommandType commandType() const override {
@@ -19,6 +19,8 @@ public:
 
 private:
     bool shouldReplan(const SinglePlanRequest& planRequest) const;
+    RRTPlanner rrtPlanner;
+
 };
 
 }  // namespace Planning
